Don't require cache directory

+++ service-scripts/p3x-load-app-specs.pl
@@ -15,16 +15,19 @@ use JSON::XS;
 use File::Slurp;
 use File::Copy::Recursive qw(dircopy);
 
-my($opt, $usage) = describe_options("%c %o container.sif cache-dir",
+my($opt, $usage) = describe_options("%c %o container.sif [cache-dir]",
 				    ["help|h" => "Show this help message."]);
 print($usage->text), exit 0 if $opt->help;
-die($usage->text) if @ARGV != 2;
+die($usage->text) if @ARGV != 1 && @ARGV != 2;
 
 my $container = shift;
 my $cache_dir = shift;
 
 -f $container or die "Container $container not found\n";
--d $cache_dir or die "Cache directory $cache_dir not found\n";
+if ($cache_dir)
+{
+    -d $cache_dir or die "Cache directory $cache_dir not found\n";
+}
 
 my $tmpdir = File::Temp->newdir();
 
@@ -42,12 +45,15 @@ my $release = $labels->{release};
 print STDERR "Loading release $release\n";
 
 my $release_name = "release-$release";
-my $release_dir = "$cache_dir/$release_name";
--d $release_dir and die "Release $release already loaded in $release_dir\n";
-
-dircopy($tmpdir, $release_dir);
-
-my $specs = Bio::KBase::AppService::AppSpecs->new("$release_dir/app_specs");
+if ($cache_dir)
+{
+    my $release_dir = "$cache_dir/$release_name";
+    -d $release_dir and die "Release $release already loaded in $release_dir\n";
+    
+    dircopy($tmpdir, $release_dir);
+}
+system("ls -lR $tmpdir/app_specs");
+my $specs = Bio::KBase::AppService::AppSpecs->new("$tmpdir/app_specs");
 print STDERR "Loading specs:\n";
 print "\t$_->{id}\n" foreach $specs->enumerate;
 
@@ -58,6 +64,9 @@ $sched->load_apps();
 # Update the "latest" symlink
 #
 
-my $latest = "$cache_dir/latest";
-unlink($latest);
-symlink($release_name, "$cache_dir/latest");
+if ($cache_dir)
+{
+    my $latest = "$cache_dir/latest";
+    unlink($latest);
+    symlink($release_name, "$cache_dir/latest");
+}
